+++ sync-for-payables/RELEASES.md
@@ -129,3 +129,13 @@ Based on:
 - [csharp v5.0.1] sync-for-payables
 ### Releases
 - [NuGet v5.0.1] https://www.nuget.org/packages/Codat.Sync.Payables/5.0.1 - sync-for-payables
+
+## 2024-10-29 13:25:05
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.423.1 (2.442.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v5.0.2] sync-for-payables
+### Releases
+- [NuGet v5.0.2] https://www.nuget.org/packages/Codat.Sync.Payables/5.0.2 - sync-for-payables